Move back from folders to root
